Update Zabbix API auth process to 6.4 and over
Starting from Zabbix 6.4, the method of passing the authorization token
via the JSON parameter "auth" is deprecated. Instead, it is now passed
via the HTTP "Authorization" header, with the authorization token itself
preceded by the word "Bearer". Continuing to pass the authorization
token as a JSON parameter with API versions 6.4 or greater results in
deprecation warnings.